Add test for binary search (toolbox 4th week) task

+++ courses/toolbox/divide_and_conquer/binary_search.h
@@ -2,7 +2,7 @@
 
 #include <vector>
 
-long binary_search(const std::vector<int> &a, int x);
+int binary_search(const std::vector<int> &a, int x);
 
 int linear_search(const std::vector<int> &a, int x);
